+++ resources/lang/de/Widget.properties
@@ -424,6 +424,7 @@ orderReturnHistoryPage = "Retouren pro Seite"
 orderReturnHistoryPageTooltip = "Anzahl der angezeigten Retouren pro Seite wählen."
 orderReturnHistoryItem = "Artikel pro Retoure"
 orderReturnHistoryItemTooltip = "Anzahl der angezeigten Artikel pro Retoure wählen."
+orderReturnHistoryMockPaymentMethod = "Zahlungsart :i"
 ; Order history
 orderHistoryLabel = "Auftragshistorie"
 orderHistoryHintTextLabel = "Hinweistext"
