+++ resources/lang/en/Widget.properties
@@ -424,6 +424,7 @@ orderReturnHistoryPage = "Returns per page"
 orderReturnHistoryPageTooltip = "Select the number of returns per page."		
 orderReturnHistoryItem = "Items per return"		
 orderReturnHistoryItemTooltip = "select the number of items per return."
+orderReturnHistoryMockPaymentMethod = "Payment method :i"
 ; Order history
 orderHistoryLabel = "Auftragshistorie"
 orderHistoryHintTextLabel = "Hinweistext"
