+++ src/Widgets/MyAccount/OrderHistoryWidget.php
@@ -10,4 +10,30 @@ class OrderHistoryWidget extends OrderHistoryBaseWidget
     {
         return $widgetSettings["ordersPerPage"]["mobile"] ?? 5;
     }
+
+    protected function getTemplateData($widgetSettings, $isPreview)
+    {
+        $previewData = null;
+
+        if ( $isPreview )
+        {
+            $previewData = $this->mockPaginatedResult(
+                function($i)
+                {
+                    return [
+                        "id"            => $i,
+                        "total"         => rand(100, 100000) / 100,
+                        "status"        => $this->getRandomStatusName(),
+                        "creationDate"  => date("Y-m-d H:i:s")
+                    ];
+                },
+                $widgetSettings["ordersPerPage"]["mobile"] ?? 5
+
+            );
+        }
+
+        return [
+            "previewData" => $previewData
+        ];
+    }
 }
